Mcc 444/Quote Layout Changes

+++ app/Repositories/QuoteFile/QuoteFileRepository.php
@@ -2,46 +2,31 @@
 
 namespace App\Repositories\QuoteFile;
 
-use App\Contracts\Repositories\QuoteFile\{
-    QuoteFileRepositoryInterface,
-    ImportableColumnRepositoryInterface as ImportableColumnRepository,
-    FileFormatRepositoryInterface as FileFormatRepository
-};
-use App\Models\{
-    Quote\BaseQuote as Quote,
-    QuoteFile\QuoteFile,
-    QuoteFile\DataSelectSeparator,
-    QuoteFile\ImportedRow
-};
+use App\Contracts\Repositories\QuoteFile\{QuoteFileRepositoryInterface};
+use App\Enum\Lock;
+use App\Models\{QuoteFile\DataSelectSeparator, QuoteFile\ImportableColumn, QuoteFile\ImportedRow, QuoteFile\QuoteFile};
 use App\Services\PdfParser\PdfOptions;
-use Illuminate\Http\UploadedFile;
-use Illuminate\Database\Eloquent\Collection;
-use Illuminate\Support\Arr;
-use Illuminate\Support\Collection as SupportCollection;
-use Storage, Str, File, DB;
+use Illuminate\Database\Eloquent\Collection as DbCollection;
+use Illuminate\Database\Query\Builder as BaseBuilder;
+use Illuminate\Support\{Arr, Collection, Facades\Cache, Facades\DB, Facades\File, Facades\Storage, Str};
+use Throwable;
+use Webpatser\Uuid\Uuid;
 
 class QuoteFileRepository implements QuoteFileRepositoryInterface
 {
     protected QuoteFile $quoteFile;
 
-    protected FileFormatRepository $fileFormat;
-
     protected DataSelectSeparator $dataSelectSeparator;
 
-    protected ImportableColumnRepository $importableColumn;
-
-    protected static ?Collection $importableColumnsCache = null;
+    protected static ?DbCollection $importableColumnsCache = null;
 
     public function __construct(
-        QuoteFile $quoteFile,
-        FileFormatRepository $fileFormat,
-        DataSelectSeparator $dataSelectSeparator,
-        ImportableColumnRepository $importableColumn
-    ) {
+        QuoteFile           $quoteFile,
+        DataSelectSeparator $dataSelectSeparator
+    )
+    {
         $this->quoteFile = $quoteFile;
-        $this->fileFormat = $fileFormat;
         $this->dataSelectSeparator = $dataSelectSeparator;
-        $this->importableColumn = $importableColumn;
     }
 
     public function all()
@@ -83,33 +68,6 @@ class QuoteFileRepository implements QuoteFileRepositoryInterface
         return $quoteFile->load('dataSelectSeparator')->makeHidden('user');
     }
 
-    public function createPdf(Quote $quote, array $attributes)
-    {
-        if (!isset($attributes['original_file_path']) || !isset($attributes['filename'])) {
-            return null;
-        }
-
-        $original_file_path = $attributes['original_file_path'];
-        $quote_file = new UploadedFile(Storage::path($attributes['original_file_path']), $attributes['filename']);
-        $format = $this->fileFormat->whereInExtension(['pdf']);
-        $quote_id = $quote->id;
-        $file_type = 'Generated PDF';
-
-        $quote->generatedPdf()->delete();
-
-        $attributes = compact('quote_file', 'format', 'file_type', 'original_file_path', 'quote_id');
-
-        if (app()->runningInConsole()) {
-            $attributes = array_merge($attributes, ['user' => $quote->user]);
-        }
-
-        $quoteFile = $this->create($attributes);
-
-        $quote->load('generatedPdf');
-
-        return $quoteFile;
-    }
-
     public function createRawData(QuoteFile $quoteFile, array $array)
     {
         $user = $quoteFile->user;
@@ -122,7 +80,7 @@ class QuoteFileRepository implements QuoteFileRepositoryInterface
 
             $importedRawData = $quoteFile->importedRawData()->make([
                 'page' => $data['page'],
-                'file_path' => $filePath
+                'file_path' => $filePath,
             ]);
 
             $importedRawData->user()->associate($user);
@@ -140,35 +98,37 @@ class QuoteFileRepository implements QuoteFileRepositoryInterface
         return $quoteFile->importedRawData()->orderBy('page')->get();
     }
 
-    public function createRowsData(QuoteFile $quoteFile, array $array)
-    {
-        /**
-         * Delete early imported data
-         */
-        DB::transaction(
-            fn () => $quoteFile->rowsData()->forceDelete(),
-            5
-        );
-
-        $this->createImportedPages($array, $quoteFile);
-    }
-
     public function createScheduleData(QuoteFile $quoteFile, array $value)
     {
         $user = $quoteFile->user;
 
-        /**
-         * Delete early imported payment schedule data
-         */
-        $quoteFile->scheduleData()->forceDelete();
+        $lock = Cache::lock(Lock::UPDATE_QUOTE_FILE($quoteFile->getKey()), 10);
 
-        $scheduleData = $quoteFile->scheduleData()->make(compact('value'));
-        $scheduleData->user()->associate($user);
-        $scheduleData->save();
+        $lock->block(30);
 
-        $quoteFile->markAsHandled();
+        DB::beginTransaction();
 
-        return $scheduleData;
+        try {
+            if ($quoteFile->scheduleData()->exists()) {
+                $quoteFile->scheduleData()->forceDelete();
+            }
+
+            $scheduleData = $quoteFile->scheduleData()->make(compact('value'));
+            $scheduleData->user()->associate($user);
+            $scheduleData->save();
+
+            $quoteFile->markAsHandled();
+
+            DB::commit();
+
+            return $scheduleData;
+        } catch (Throwable $e) {
+            DB::rollBack();
+
+            throw $e;
+        } finally {
+            optional($lock)->release();
+        }
     }
 
     public function getRowsData(QuoteFile $quoteFile)
@@ -183,7 +143,7 @@ class QuoteFileRepository implements QuoteFileRepositoryInterface
 
     public function find(string $id)
     {
-        return $this->quoteFile->query()->whereId($id)->firstOrFail();
+        return $this->quoteFile->query()->whereKey($id)->firstOrFail();
     }
 
     public function findByClause(array $clause)
@@ -191,56 +151,39 @@ class QuoteFileRepository implements QuoteFileRepositoryInterface
         return $this->quoteFile->query()->where($clause)->first();
     }
 
+    public function findByQuote(string $quoteID, string $fileType)
+    {
+        return $this->quoteFile->query()
+            ->whereExists(fn(BaseBuilder $builder) => $builder->selectRaw('1')->from('quotes')
+                ->where('id', $quoteID)
+                ->where(fn(BaseBuilder $builder) => $builder->whereColumn('quote_files.id', 'quotes.distributor_file_id')->orWhereColumn('quote_files.id', 'quotes.schedule_file_id')))
+            ->where('file_type', $fileType)
+            ->first();
+    }
+
     public function exists(string $id)
     {
         return $this->quoteFile->whereId($id)->exists();
     }
 
-    public function deletePriceListsExcept(QuoteFile $quoteFile)
-    {
-        $exceptedQuoteFileId = $quoteFile->id;
-        $quote = $quoteFile->quote;
-
-        return $quote->quoteFiles()->priceLists()->whereKeyNot($exceptedQuoteFileId)->delete();
-    }
-
-    public function deletePaymentSchedulesExcept(QuoteFile $quoteFile)
-    {
-        $exceptedQuoteFileId = $quoteFile->id;
-        $quote = $quoteFile->quote;
-
-        return $quote->quoteFiles()->paymentSchedules()->whereKeyNot($exceptedQuoteFileId)->delete();
-    }
-
-    public function deleteExcept(QuoteFile $quoteFile)
-    {
-        if ($quoteFile->isSchedule()) {
-            return $this->deletePaymentSchedulesExcept($quoteFile);
-        }
-
-        return $this->deletePriceListsExcept($quoteFile);
-    }
-
     public function replicatePriceList(QuoteFile $quoteFile, ?string $quoteId = null): QuoteFile
     {
         $quote_file_id = $quoteFile->id;
-        $user_id = $quoteFile->user_id;
 
         /** @var QuoteFile */
         $quoteFileCopy = tap($quoteFile->replicate(), function ($file) use ($quoteId) {
-            $file->quote()->associate($quoteId);
             $file->save();
         });
 
         $new_quote_file_id = $quoteFileCopy->getKey();
 
-        $tempRowsTable = 'temp_imported_rows_table_' . uniqid();
+        $tempRowsTable = 'temp_imported_rows_table_'.uniqid();
         $importedRowColumn = 'replicated_row_id';
 
         /** Generating new Ids for Imported Columns and Rows in the temporary table. */
-        DB::select(
+        DB::statement(
             "create temporary table `{$tempRowsTable}`
-            select uuid() as `id`, `id` as {$importedRowColumn}, `quote_file_id`, `columns_data`, `page`, `is_selected`, `group_name` from `imported_rows`
+            select uuid() as `id`, `id` as {$importedRowColumn}, `quote_file_id`, `columns_data`, `page`, `is_selected` from `imported_rows`
                 where `imported_rows`.`quote_file_id` = :quote_file_id",
             compact('quote_file_id')
         );
@@ -248,15 +191,12 @@ class QuoteFileRepository implements QuoteFileRepositoryInterface
         /** We are marking temporary rows as selected if the related payload is present in the current request. */
         $this->updateTempSelectedImportedRows($tempRowsTable, $importedRowColumn);
 
-        /** We are updating temporary group description if the related payload is present in the current request. */
-        // $this->updateTempGroupDescription($tempRowsTable, $importedRowColumn);
-
         /** Inserting Imported Rows with new Ids. */
         DB::insert(
-            "insert into `imported_rows` (`id`, `{$importedRowColumn}`, `user_id`, `quote_file_id`, `columns_data`, `page`, `is_selected`, `group_name`)
-            select `id`, `{$importedRowColumn}`, :user_id, :new_quote_file_id, `columns_data`, `page`, `is_selected`, `group_name`
+            "insert into `imported_rows` (`id`, `{$importedRowColumn}`,`quote_file_id`, `columns_data`, `page`, `is_selected`)
+            select `id`, `{$importedRowColumn}`, :new_quote_file_id, `columns_data`, `page`, `is_selected`
             from `{$tempRowsTable}`",
-            compact('user_id', 'new_quote_file_id')
+            compact('new_quote_file_id')
         );
 
         return $quoteFileCopy;
@@ -264,14 +204,10 @@ class QuoteFileRepository implements QuoteFileRepositoryInterface
 
     public function resolveFileType(string $needle): ?string
     {
-        switch ($needle) {
-            case 'price':
-                return QFT_PL;
-                break;
-            case 'schedule':
-                return QFT_PS;
-                break;
-        }
+        return match ($needle) {
+            'price' =>  QFT_PL,
+            'schedule' => QFT_PS,
+        };
     }
 
     public function resolveFilepath(?QuoteFile $quoteFile): string
@@ -285,7 +221,7 @@ class QuoteFileRepository implements QuoteFileRepositoryInterface
 
     protected function createImportedPages(array $array, QuoteFile $quoteFile)
     {
-        $pages = collect($array)->filter(fn ($item) => (isset($item['page']) && isset($item['rows'])));
+        $pages = collect($array)->filter(fn($item) => (isset($item['page']) && isset($item['rows'])));
 
         $importedRows = [];
 
@@ -295,52 +231,61 @@ class QuoteFileRepository implements QuoteFileRepositoryInterface
             array_push($importedRows, ...$this->createImportedRows($rows, $quoteFile, $page));
         });
 
-        if (count($importedRows) < 1) {
-            $quoteFile->setException(QFNRF_01, 'QFNRF_01');
-            $quoteFile->markAsUnHandled();
-            $quoteFile->throwExceptionIfExists();
+        if (empty($importedRows)) {
+            return;
         }
 
-        $quoteFile->rowsData()->saveMany($importedRows);
+        $rows = array_map(fn(ImportedRow $row) => $row->getAttributes() + [
+                'id' => (string)Uuid::generate(4),
+                'created_at' => now()->toDateTimeString(),
+                'updated_at' => now()->toDateTimeString(),
+            ], $importedRows);
+
+        ImportedRow::insert($rows);
     }
 
-    protected function createImportedRows(array $rows, QuoteFile $quoteFile, int $page): SupportCollection
+    protected function createImportedRows(array $rows, QuoteFile $quoteFile, int $page): Collection
     {
         $quote_file_id = $quoteFile->id;
         $user_id = $quoteFile->user->id;
         $attributes = compact('quote_file_id', 'user_id', 'page');
 
         return collect($rows)
-            ->map(fn ($row) => ImportedRow::make(
+            ->map(fn($row) => ImportedRow::make(
                 $attributes + [
                     'columns_data' => $this->mapColumnsData($row),
-                    'is_one_pay' => (bool) Arr::get($row, PdfOptions::SYSTEM_HEADER_ONE_PAY)
+                    'is_one_pay' => (bool)Arr::get($row, PdfOptions::SYSTEM_HEADER_ONE_PAY),
                 ]
             ));
     }
 
-    protected function mapColumnsData(array $row): SupportCollection
+    protected function mapColumnsData(array $row): Collection
     {
         $importableColumns = $this->getImportableColumns();
 
         Arr::forget($row, PdfOptions::SYSTEM_HEADER_ONE_PAY);
 
-        return SupportCollection::wrap($row)->map(
-            fn ($value, $name) => [
+        return Collection::wrap($row)->map(
+            fn($value, $name) => [
                 'header' => Str::header($name),
                 'value' => $value,
-                'importable_column_id' => optional($importableColumns->get($name))->id
+                'importable_column_id' => optional($importableColumns->get($name))->id,
             ]
         )->values();
     }
 
-    protected function getImportableColumns(): Collection
+    protected function getImportableColumns(): DbCollection
     {
         if (isset(static::$importableColumnsCache)) {
             return static::$importableColumnsCache;
         }
 
-        return static::$importableColumnsCache = $this->importableColumn->allSystem()->keyBy('name');
+        return static::$importableColumnsCache = ImportableColumn::query()
+            ->orderBy('order')
+            ->where('is_system', true)
+            ->with('aliases')
+            ->get()
+            ->keyBy('name');
     }
 
     private function updateTempSelectedImportedRows(string $table, string $column): void
@@ -349,40 +294,13 @@ class QuoteFileRepository implements QuoteFileRepositoryInterface
             return;
         }
 
-        $selectedRowsIds = (array) request()->input('quote_data.selected_rows', []);
-        $reject = (bool) request()->input('quote_data.selected_rows_is_rejected', false);
+        $selectedRowsIds = (array)request()->input('quote_data.selected_rows', []);
+        $reject = (bool)request()->input('quote_data.selected_rows_is_rejected', false);
 
         $updatableScope = $reject ? 'whereNotIn' : 'whereIn';
 
         DB::table($table)->update(['is_selected' => false]);
 
         DB::table($table)->{$updatableScope}($column, $selectedRowsIds)->update(['is_selected' => true]);
-    }
-
-    private function updateTempGroupDescription(string $table, string $column): void
-    {
-        if (!request()->is('api/quotes/groups/*')) {
-            return;
-        }
-
-        $rowsIds = request()->input('rows', []);
-        $name = request()->input('name');
-
-        /** Create Group Description request */
-        if (request()->isMethod('post')) {
-            DB::table($table)->whereIn($column, $rowsIds)->update(['group_name' => $name]);
-        }
-
-        /** Update Group Description request */
-        if (request()->isMethod('patch')) {
-            DB::table($table)->where('group_name', request()->group_name)->update(['group_name' => null]);
-            DB::table($table)->whereIn($column, $rowsIds)->update(['group_name' => $name]);
-        }
-
-        /** Move Group Description rows request */
-        if (request()->isMethod('put')) {
-            DB::table($table)->where('group_name', request()->from_group_name)->whereIn($column, $rowsIds)
-                ->update(['group_name' => request()->to_group_name]);
-        }
     }
 }
